refactor mobileappcomingsoon, add note about globals
