+++ src/app/globals.css
@@ -1,3 +1,8 @@
+/* The globals.css file is used to define variables used across the application, as well as animations. 
+* It should not contain any styles that are specific to a component or page.
+* Do not create any custom classes or media queries here.
+*/
+
 @import 'tailwindcss';
 @plugin "tailwindcss-animate";
 
@@ -2682,14 +2687,6 @@ mux-player {
 			-moz-osx-font-smoothing: grayscale;
 			text-rendering: optimizeLegibility;
 		}
-	}
-}
-
-/* Safari-specific mobile app coming soon page styles */
-@supports (-webkit-touch-callout: none) {
-	/* iOS Safari specific styles */
-	.fixed.inset-0 {
-		min-height: -webkit-fill-available;
 	}
 }
