--- src/components/molecules/MobileAppComingSoon/index.ts
@@ -1,1 +1,0 @@
-export { MobileAppComingSoon } from './MobileAppComingSoon';
